If vcvarsall is not set get it from environment variable.
If vcvarsall is not set manually in the config file, bob will look for
it using the VCINSTALLDIR env variable, this requires a Visual Studio 14
or more.

If vcvarsall was not set in cfg file and bob failes to get it from the
env variable a message will be printed to the console.